Add more fields, add date fields format, bug fixes
