+++ src/utils/messages.json
@@ -90,13 +90,16 @@
   "candidate.info.phone" : "Phone",
   "candidate.info.skype" : "Skype",
   "candidate.info.email" : "Email",
+  "candidate.info.education" : "Education",
+  "candidate.info.professionalCertificates" : "Certificates",
   "candidate.info.bestContactTime" : "Best contact time",
   "candidate.info.stackType" : "Stack",
   "candidate.info.englishLevelName" : "English level",
   "candidate.info.registrationDate" : "Registration date",
   "candidate.info.otherInfo" : "Other info",
   "candidate.info.testTaskEvaluation" : "Test task evaluation",
-  "candidate.info.currentJob" : "Soft skills",
+  "candidate.info.links" : "Links",
+  "candidate.info.currentJob" : "Current job",
   "candidate.button.edit" : "Edit",
   "addprogram.title": "Add Program",
   "addprogram.button.add": "Add",
